Time: 1 ms (87.41%), Space: 42.5 MB (87.55%) - LeetHub

+++ 3870-count-commas-in-range/3870-count-commas-in-range.java
@@ -2,9 +2,9 @@ class Solution
 {
     public int countCommas(int n) 
     {
-        long ans =0;
+        int ans =0;
         for(long i=1000;i<=n;i*=1000)
             ans += (n-i+1);
-        return (int)ans;
+        return ans;
     }
 }
